feat: rename recoil folder to recoil-store
