+++ pages/index.tsx
@@ -3,7 +3,7 @@ import { useWallet } from '@solana/wallet-adapter-react'
 import { CandyMachineCard, Spinner, Title } from 'components'
 import { useRPC } from 'hooks'
 import { CANDY_MACHINE_PROGRAM_V2_ID } from 'lib/candy-machine/constants'
-import { candyMachinesState } from 'lib/recoil/atoms'
+import { candyMachinesState } from 'lib/recoil-store/atoms'
 import type { NextPage } from 'next'
 import Head from 'next/head'
 import Link from 'next/link'
